+++ apps/api/src/modules/notifications/scheduler.module.ts
@@ -1,12 +1,11 @@
 import { Module } from '@nestjs/common';
 import { SchedulerService } from './scheduler.service';
 import { NotificationsModule } from './notifications.module';
-import { OverdueModule } from '../overdue/overdue.module';
 import { ReorderPointsModule } from '../reorder-points/reorder-points.module';
 import { LineOaModule } from '../line-oa/line-oa.module';
 
 @Module({
-  imports: [NotificationsModule, OverdueModule, ReorderPointsModule, LineOaModule],
+  imports: [NotificationsModule, ReorderPointsModule, LineOaModule],
   providers: [SchedulerService],
 })
 export class SchedulerModule {}
